moving to OOP
